+++ spec/identity_registration_spec.rb
@@ -47,6 +47,90 @@ describe 'openstack-network::identity_registration' do
         )
       end
 
+      it 'with different admin url values' do
+        admin_url = 'https://admin.host:123/admin_path'
+        general_url = 'http://general.host:456/general_path'
+
+        # Set the general endpoint
+        node.set['openstack']['endpoints']['network-api']['uri'] = general_url
+        # Set the admin endpoint override
+        node.set['openstack']['endpoints']['admin']['network-api']['uri'] = admin_url
+        expect(chef_run).to create_endpoint_openstack_identity_register(
+          'Register Network Endpoint'
+        ).with(
+          auth_uri: 'http://127.0.0.1:35357/v2.0',
+          bootstrap_token: 'bootstrap-token',
+          service_type: 'network',
+          endpoint_region: 'RegionOne',
+          endpoint_adminurl: admin_url,
+          endpoint_internalurl: general_url,
+          endpoint_publicurl: general_url
+        )
+      end
+
+      it 'with different public url values' do
+        public_url = 'https://public.host:789/public_path'
+        general_url = 'http://general.host:456/general_path'
+
+        # Set the general endpoint
+        node.set['openstack']['endpoints']['network-api']['uri'] = general_url
+        # Set the public endpoint override
+        node.set['openstack']['endpoints']['public']['network-api']['uri'] = public_url
+        expect(chef_run).to create_endpoint_openstack_identity_register(
+          'Register Network Endpoint'
+        ).with(
+          auth_uri: 'http://127.0.0.1:35357/v2.0',
+          bootstrap_token: 'bootstrap-token',
+          service_type: 'network',
+          endpoint_region: 'RegionOne',
+          endpoint_adminurl: general_url,
+          endpoint_internalurl: general_url,
+          endpoint_publicurl: public_url
+        )
+      end
+
+      it 'with different internal url values' do
+        internal_url = 'http://internal.host:456/internal_path'
+        general_url = 'http://general.host:456/general_path'
+
+        # Set the general endpoint
+        node.set['openstack']['endpoints']['network-api']['uri'] = general_url
+        # Set the internal endpoint override
+        node.set['openstack']['endpoints']['internal']['network-api']['uri'] = internal_url
+        expect(chef_run).to create_endpoint_openstack_identity_register(
+          'Register Network Endpoint'
+        ).with(
+          auth_uri: 'http://127.0.0.1:35357/v2.0',
+          bootstrap_token: 'bootstrap-token',
+          service_type: 'network',
+          endpoint_region: 'RegionOne',
+          endpoint_adminurl: general_url,
+          endpoint_internalurl: internal_url,
+          endpoint_publicurl: general_url
+        )
+      end
+
+      it 'with different internal,public, and admin url values' do
+        admin_url = 'https://admin.host:123/admin_path'
+        internal_url = 'http://internal.host:456/internal_path'
+        public_url = 'https://public.host:789/public_path'
+
+        node.set['openstack']['endpoints']['internal']['network-api']['uri'] = internal_url
+        node.set['openstack']['endpoints']['public']['network-api']['uri'] = public_url
+        node.set['openstack']['endpoints']['admin']['network-api']['uri'] = admin_url
+
+        expect(chef_run).to create_endpoint_openstack_identity_register(
+          'Register Network Endpoint'
+        ).with(
+          auth_uri: 'http://127.0.0.1:35357/v2.0',
+          bootstrap_token: 'bootstrap-token',
+          service_type: 'network',
+          endpoint_region: 'RegionOne',
+          endpoint_adminurl: admin_url,
+          endpoint_internalurl: internal_url,
+          endpoint_publicurl: public_url
+        )
+      end
       it 'with custom region override' do
         node.set['openstack']['network']['region'] = 'netRegion'