+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This file is used to list changes made in each version of cookbook-openstack-net
 
 ## 10.1.1
 * Removed all files not needed by the neutron cookbook
+* Add radvd to neutron_l3_packages to support ipv6
 
 ## 10.1.0
 * Add support for Neutron VPN Service