+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This file is used to list changes made in each version of cookbook-openstack-net
 * Add radvd to neutron_l3_packages to support ipv6
 * Use common specific_endpoint routines (bug 1412919)
 * Parameterize dhcp_agents_per_network
+* Parameterize neutron l3 agent HA parameters
 
 ## 10.1.0
 * Add support for Neutron VPN Service